Default(T) Protection: In both the Observe and CollapseWeighted methods (and similarly in the Eigenstates version), we now check whether the picked (or resulting) value is the undesired default using our optional value validator. This validator is by default set to v => !EqualityComparer<T>.Default.Equals(v, default), but you can supply your own when constructing a QuBit<T>.
Runtime Flag:
The static class QuantumConfig holds the flag ForbidDefaultOnCollapse. If true (the default), then if a collapse produces a default value, an InvalidOperationException is thrown. This makes the behavior configurable.

Immutable State Post-Collapse:
We introduce an EnsureMutable() helper that throws an exception if a mutation is attempted after collapse (i.e. when _isActuallyCollapsed is true). All mutators (for example, Append()) now call this guard to enforce immutability.
The CloneMutable() method creates a deep copy of the current qubit (including the list and weights) but does not carry over the collapsed state, letting you “reset” and modify the clone.